Add unit test for MLT renderer.
