+++ test/CMakeLists.txt
@@ -39,10 +39,10 @@ if (${SPICA_BUILD_TEST} STREQUAL "ON")
       test_sampler.cc
       test_trimesh.cc
       test_kdtree.cc
-      test_renderer.cc
       test_scene.cc
       test_random.cc
-      test_pt_renderer.cc)
+      test_renderer_helper.cc
+      test_renderer.cc)
 
   add_executable(${TEST_NAME} ${SOURCE_FILES})
   target_link_libraries(${TEST_NAME} ${GTEST_LIBRARY})
